[IMP] mail: unskip "[technical] sidebar: cannot unpin channel group_based_subscription: mandatorily pinned"
